backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.27.0
pandas==2.2.2
python-dotenv==1.0.1
cachetools==5.3.3
fredapi==0.5.1
gunicorn==23.0.0
